\name{showme}
\alias{showme} 
\title{Display Abbreviated Output of FORSEARCH_xxx Function 
}
\description{Output of forsearch_xxx function can be voluminous. This function displays the 
output in an abbreviated format. Primarily for programmer use.
}

\usage{showme(x, verbose = TRUE)
}
\arguments{
  \item{x}{Name of forsearch_xxx output file}
  \item{verbose}{If TRUE, indicates the beginning and end of function run}
}
\value{Abbreviated printout of output of forsearch_lm function
}
\author{William R. Fairweather 
}
 \keyword{ manip }
